# Conflicts:
#	cmake/scripts/FindVersion.cmake
#	src/ai/ai.hpp
#	src/ai/ai_core.cpp
#	src/autocompletion.h
#	src/autoreplace_gui.cpp
#	src/base_station_base.h
#	src/build_vehicle_gui.cpp
#	src/console_cmds.cpp
#	src/engine_gui.h
#	src/game/game.hpp
#	src/game/game_core.cpp
#	src/gamelog_internal.h
#	src/graph_gui.cpp
#	src/industry_gui.cpp
#	src/network/network_client.cpp
#	src/network/network_server.cpp
#	src/newgrf/newgrf_act3.cpp
#	src/newgrf_commons.h
#	src/newgrf_spritegroup.cpp
#	src/newgrf_spritegroup.h
#	src/newgrf_station.cpp
#	src/order_backup.h
#	src/os/macosx/crashlog_osx.cpp
#	src/os/unix/crashlog_unix.cpp
#	src/os/windows/crashlog_win.cpp
#	src/picker_gui.h
#	src/rail_map.h
#	src/road_map.h
#	src/safeguards.h
#	src/saveload/station_sl.cpp
#	src/screenshot_type.h
#	src/script/squirrel_helper.hpp
#	src/settingentry_gui.h
#	src/station_base.h
#	src/station_cmd.cpp
#	src/station_gui.cpp
#	src/vehicle_gui.cpp
#	src/video/video_driver.hpp
#	src/water_cmd.cpp
#	src/water_map.h
#	src/widget_type.h

+++ src/gfx.cpp
@@ -624,7 +624,7 @@ static int DrawLayoutLine(const ParagraphLayouter::Line &line, int y, int left, 
 		int dpi_right = dpi->left + dpi->width - 1;
 		TextColour last_colour = initial_colour;
 
-		for (int run_index = 0; run_index < line.CountRuns(); run_index++) {
+		for (size_t run_index = 0; run_index < line.CountRuns(); run_index++) {
 			const ParagraphLayouter::VisualRun &run = line.GetVisualRun(run_index);
 			const auto &glyphs = run.GetGlyphs();
 			const auto &positions = run.GetPositions();
@@ -639,7 +639,7 @@ static int DrawLayoutLine(const ParagraphLayouter::Line &line, int y, int left, 
 			if (do_shadow && (!fc->GetDrawGlyphShadow() || !colour_has_shadow)) continue;
 			ctx.SetColourRemap(do_shadow ? TC_BLACK : colour); // the last run also sets the colour for the truncation dots
 
-			for (int i = 0; i < run.GetGlyphCount(); i++) {
+			for (size_t i = 0; i < run.GetGlyphCount(); i++) {
 				GlyphID glyph = glyphs[i];
 
 				/* Not a valid glyph (empty) */
